Improve Space friend invitations and avatar contrast
