+++ web/apps/space/src/pages/app/index.tsx
@@ -1,3 +1,4 @@
+import { SpaceAddFriendDialog } from "components/AddFriendDialog";
 import { SpaceFriendLimitToast } from "components/FriendLimitToast";
 import { SpaceFriendRequestCanceledToast } from "components/FriendRequestCanceledToast";
 import { SpacePageMeta } from "components/PageMeta";
@@ -27,6 +28,7 @@ import {
     loadCurrentSpacePostAssetURL,
     loadCurrentUnreadStatus,
     replyToCurrentPost,
+    requestFriendByUsername,
     sendCurrentPoke,
     setCurrentPostLiked,
     type SpaceFriendRequest,
@@ -54,6 +56,7 @@ const Page: React.FC = () => {
     } = useSpaceAppState();
     const [friendRequestSentToastName, setFriendRequestSentToastName] =
         useState<string>();
+    const [isAddFriendOpen, setIsAddFriendOpen] = useState(false);
     const [friendRequests, setFriendRequests] = useState<SpaceFriendRequest[]>(
         [],
     );
@@ -269,11 +272,6 @@ const Page: React.FC = () => {
                 isFriendRequestsLoading={isFriendRequestsLoading}
                 isHomeCacheLoading={isHomeCacheLoading}
                 profile={profile}
-                profileLink={
-                    profile
-                        ? spaceInviteURL({ spaceUsername: profile.username })
-                        : undefined
-                }
                 viewerSpaceId={spaceId ?? profile?.spaceId}
                 showInstallPrompt={
                     profileLoadStatus == "ready" &&
@@ -282,6 +280,7 @@ const Page: React.FC = () => {
                     !isFriendsLoading
                 }
                 onFriendRequestSentToastClose={closeFriendRequestSentToast}
+                onAddFriend={() => setIsAddFriendOpen(true)}
                 onAcceptFriendRequest={async (requestID) => {
                     if (!profile?.spaceId) throw new Error("Missing space.");
                     const sentRequestCount = friendRequests.filter(
@@ -455,6 +454,62 @@ const Page: React.FC = () => {
             {showFriendRequestCanceledToast && (
                 <SpaceFriendRequestCanceledToast
                     onClose={() => setShowFriendRequestCanceledToast(false)}
+                />
+            )}
+            {profile && (
+                <SpaceAddFriendDialog
+                    friendRequests={friendRequests}
+                    friends={friends}
+                    open={isAddFriendOpen}
+                    onClose={() => setIsAddFriendOpen(false)}
+                    profileLink={spaceInviteURL({
+                        spaceUsername: profile.username,
+                    })}
+                    username={profile.username}
+                    onAddFriend={async (username) => {
+                        const actorSpaceId = profile.spaceId;
+                        if (!actorSpaceId) throw new Error("Missing space.");
+                        const status = await requestFriendByUsername({
+                            spaceUsername: username,
+                        });
+                        try {
+                            if (status == "friend") {
+                                clearSpaceFriendsCache();
+                                const [requests, friends] = await Promise.all([
+                                    loadCurrentFriendRequests(actorSpaceId),
+                                    loadCurrentSpaceFriends(actorSpaceId),
+                                ]);
+                                setFriendRequests(requests);
+                                setFriends(friends);
+                                const refreshedHomePosts =
+                                    await refreshSpaceHomePosts(
+                                        actorSpaceId,
+                                        friends,
+                                    );
+                                if (refreshedHomePosts) {
+                                    setLatestPosts(
+                                        refreshedHomePosts.latestPosts,
+                                    );
+                                    setUnreadPosts(
+                                        refreshedHomePosts.unreadPosts,
+                                    );
+                                }
+                                await refreshUnreadStatus(actorSpaceId);
+                            } else {
+                                setFriendRequests(
+                                    await loadCurrentFriendRequests(
+                                        actorSpaceId,
+                                    ),
+                                );
+                            }
+                        } catch (error) {
+                            log.error(
+                                "Failed to refresh friends after sending request",
+                                error,
+                            );
+                        }
+                        return status;
+                    }}
                 />
             )}
             {showFriendLimitToast && (
